use a statistics store #377
